* Fixed: Compilation warnings with -Woverloaded-virtual switch.
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@7740 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/include/editorbase.h
@@ -289,7 +289,7 @@ class DLLIMPORT EditorBase : public wxPanel
           * @param id An event handler's ID.
           * @return The created submenu or NULL if not applicable.
           */
-        virtual wxMenu* CreateContextSubMenu(int id); // For context menus
+        virtual wxMenu* CreateContextSubMenu(long id); // For context menus
 
         /** Creates context menu items, both before and after creating plugins menu items.
           * @param popup The popup menu.
